Edited all Long/long to Integer/int. Long got edited out from DAO classes and long from entities

+++ src/main/java/app/dao/ActorDAO.java
@@ -4,7 +4,7 @@ import app.entities.Actor;
 import app.entities.Movie;
 import jakarta.persistence.EntityManagerFactory;
 
-public class ActorDAO extends GenericDAO<Actor, Long> {
+public class ActorDAO extends GenericDAO<Actor, Integer> {
     public ActorDAO(EntityManagerFactory emf) {
         super(emf, Actor.class);
     }
